product Detail page Done

+++ backend/db/migrations/20220609035410-create-product.js
@@ -14,7 +14,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(20, 2),
       },
       imageUrl: {
         allowNull: false,
